fix: specify and pass on secrets explicitly
It seems the behavior in GitHub Actions has changed.
Previously it was possible to reuse a workflow from another organization
which could in turn reuse workflow from that same organization while
inheriting secrets using `secrets: inherit`.
This currently is no longer possible. Secrets have to be specified
explicitly for both kinds of reuses where before it was only necessary
for the workflow that was explicitly referenced from another
organization.